refactor(ui): extract model data from ModelPicker.tsx into model-data.ts (#94)
* docs: refresh codebase map after v0.8.0 release

Refresh all 7 codebase documentation files to reflect the current state after Round 5 architecture extractions (tool-status, SyntaxHighlighter, plan-converter, help-text, Login Flow Controller), the stable prompt prefix fix, and the v0.8.0 release.

* refactor(ui): extract model data from ModelPicker.tsx into model-data.ts

Round 6 Candidate #1 — extract static model data (PROVIDER_MODELS with 80 model entries) and helper functions from ModelPicker.tsx (380 lines) into its own module model-data.ts (~130 lines).

- model-data.ts (new): PROVIDER_MODELS, PROVIDER_NAMES, getModelsForProviders, getEnabledProviders, getProviderDisplayName, DEFAULT_MODELS + type exports (EnabledProviders, ModelPickerItem, ProviderModels)
- ModelPicker.tsx (modified, -195 lines): keeps only the ModelPicker component + UI helpers (groupModelsByProvider, buildVisibleItems). Re-exports moved types/values from model-data.ts for backward compatibility.
- All 7 importing files continue to work via re-export chain. Lint clean.
